test: consolidate LoginLink E2E and fix logout navigation (#1076)
Remove duplicate LoginLinkHeaderTests; fix EnsureLoggedOut to follow
logout to /login then return to /. Harden animation assertion using
box-shadow delta. Revert redundant AcceptanceTestBase hooks superseded
by RunWithExtraBrowserContextAsync.
