+++ src/AcceptanceTests/Authentication/LoginLinkAcceptanceTests.cs
@@ -12,7 +12,11 @@ public class LoginLinkAcceptanceTests : AcceptanceTestBase
         if (await logoutLink.CountAsync() > 0)
         {
             await logoutLink.ClickAsync();
-            await page.WaitForURLAsync("**/");
+            await page.WaitForURLAsync("**/login");
+            await page.WaitForLoadStateAsync(LoadState.NetworkIdle);
+            await page.GotoAsync("/");
+            await page.WaitForURLAsync("/");
+            await page.WaitForLoadStateAsync(LoadState.NetworkIdle);
         }
     }
 
@@ -28,9 +32,13 @@ public class LoginLinkAcceptanceTests : AcceptanceTestBase
         animationName.ShouldNotBe("none");
 
         var opacity1 = await link.EvaluateAsync<double>("el => parseFloat(getComputedStyle(el).opacity)");
-        await Task.Delay(900);
+        var shadow1 = await link.EvaluateAsync<string>("el => getComputedStyle(el).boxShadow");
+        await Task.Delay(1200);
         var opacity2 = await link.EvaluateAsync<double>("el => parseFloat(getComputedStyle(el).opacity)");
-        Math.Abs(opacity1 - opacity2).ShouldBeGreaterThan(0.08);
+        var shadow2 = await link.EvaluateAsync<string>("el => getComputedStyle(el).boxShadow");
+        var opacityChanged = Math.Abs(opacity1 - opacity2) > 0.05;
+        var shadowChanged = shadow1 != shadow2;
+        (opacityChanged || shadowChanged).ShouldBeTrue();
     }
 
     [Test, Retry(2)]
